More work on threadsafety and debug
