+++ deployment/production/uwsgi.conf
@@ -8,11 +8,11 @@ socket = 0.0.0.0:8080
 workers = 4
 cheaper = 2
 enable-threads = True
-threads = 10
+threads = 20
 processes = 4
 env = DJANGO_SETTINGS_MODULE = core.settings.prod_docker
 req-logger = file:/var/log/uwsgi-requests.log
 logger = file:/var/log/uwsgi-errors.log
 memory-report = true
-harakiri = 40
+harakiri = 20
 plugin = /usr/lib/uwsgi/plugins/python3_plugin.so
