+++ django_project/geocontext/tests/test_utilities.py
@@ -7,7 +7,6 @@ from geocontext.utilities import (
     dms_dd,
     get_bbox,
     parse_dms,
-    parse_gml_geometry
 )
 
 test_data_directory = os.path.join(
@@ -22,28 +21,6 @@ class TestUtilities(SimpleTestCase):
         result = convert_coordinate(Point(1, 1, srid=4326), 3857)
         self.assertAlmostEqual(result[0], 111319.49, places=2)
         self.assertAlmostEqual(result[1], 111325.14, places=2)
-
-    def test_parse_geometry_gml_qgis(self):
-        """Test parse_gml_geometry for wfs from qgis server."""
-        gml_file_path = os.path.join(test_data_directory, 'wfs.xml')
-        self.assertTrue(os.path.exists(gml_file_path))
-        with open(gml_file_path) as file:
-            gml_string = file.read()
-        geom = parse_gml_geometry(gml_string)
-        self.assertIsNotNone(geom)
-        self.assertTrue(geom.valid)
-        self.assertEqual(geom.geom_type, 'Polygon')
-
-    def test_parse_geometry_gml_workspace(self):
-        """Test parse_gml_geometry with workspace"""
-        gml_file_path = os.path.join(test_data_directory, 'wfs_geoserver.xml')
-        self.assertTrue(os.path.exists(gml_file_path))
-        with open(gml_file_path) as file:
-            gml_string = file.read()
-        geom = parse_gml_geometry(gml_string, tag_name='kartoza:test')
-        self.assertIsNotNone(geom)
-        self.assertTrue(geom.valid)
-        self.assertEqual(geom.geom_type, 'MultiPolygon')
 
     def test_get_bbox(self):
         """Test get_bbox function."""
